# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Bank API
  description: Bankcard API made as a test task
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Generated server url
security:
  - bearerAuth: []
...
paths:
...
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Register a user
      description: Registers user with given credentials and send it back with jwt-token
      operationId: register
      requestBody:
        description: User creation parameters
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegisterDto"
        required: true
      responses:
        "201":
          description: User was created and returned along with jwt token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponseDto"
        "403":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Login a user
      description: Logs user in by given credentials and send it back with jwt-token
      operationId: login
      requestBody:
        description: User login parameters
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLoginDto"
        required: true
      responses:
        "200":
          description: User was logged in and returned along with jwt token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponseDto"
        "404":
          description: User with given credentials not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/user/update/password:
    patch:
      tags:
        - User - Self
      summary: Change password
      description: Changes user's password and checks the old given one
      operationId: changePassword
      requestBody:
        description: Old and new password
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLoginDto"
        required: true
      responses:
        "204":
          description: Password was changed
        "401":
          description: User is not authorized for the action
        "403":
          description: Given old password is invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/user/update/nickname:
    patch:
      tags:
        - User - Self
      summary: Change nickname
      description: Changes user's nickname
      operationId: changeNickname
      requestBody:
        description: New nickname
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLoginDto"
        required: true
      responses:
        "200":
          description: Nickname was changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDto"
        "401":
          description: User is not authorized for the action
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/user/update/email:
    patch:
      tags:
        - User - Self
      summary: Change email
      description: Changes user's email and checks it for uniqueness
      operationId: changeEmail
      requestBody:
        description: New email
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLoginDto"
        required: true
      responses:
        "200":
          description: Email was changed. Auth response was returned to refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponseDto"
        "401":
          description: User is not authorized for the action
        "403":
          description: Email is already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/user:
    delete:
      tags:
        - User - Self
      summary: Delete account
      description: Deletes user's account
      operationId: deleteUser
      responses:
        "204":
          description: User was deleted
        "401":
          description: User is not authorized for the action
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ContentDisposition:
      type: object
      properties:
        type:
          type: string
        name:
          type: string
        filename:
          type: string
        charset:
          type: string
        size:
          type: integer
          format: int64
          deprecated: true
        creationDate:
          type: string
          format: date-time
          deprecated: true
        modificationDate:
          type: string
          format: date-time
          deprecated: true
        readDate:
          type: string
          format: date-time
          deprecated: true
        attachment:
          type: boolean
        formData:
          type: boolean
        inline:
          type: boolean
    ErrorResponse:
      type: object
      properties:
        body:
          $ref: "#/components/schemas/ProblemDetail"
        headers:
          type: object
          properties:
            connection:
              type: array
              items:
                type: string
            contentDisposition:
              $ref: "#/components/schemas/ContentDisposition"
            acceptCharset:
              type: array
              items:
                type: string
            location:
              type: string
              format: uri
            empty:
              type: boolean
            host:
              type: object
              properties:
                hostString:
                  type: string
                address:
                  type: object
                  properties:
                    hostAddress:
                      type: string
                    address:
                      type: array
                      items:
                        type: string
                        format: byte
                    hostName:
                      type: string
                    linkLocalAddress:
                      type: boolean
                    multicastAddress:
                      type: boolean
                    anyLocalAddress:
                      type: boolean
                    loopbackAddress:
                      type: boolean
                    siteLocalAddress:
                      type: boolean
                    mcglobal:
                      type: boolean
                    mcnodeLocal:
                      type: boolean
                    mclinkLocal:
                      type: boolean
                    mcsiteLocal:
                      type: boolean
                    mcorgLocal:
                      type: boolean
                    canonicalHostName:
                      type: string
                port:
                  type: integer
                  format: int32
                unresolved:
                  type: boolean
                hostName:
                  type: string
            all:
              type: object
              additionalProperties:
                type: string
              writeOnly: true
            lastModified:
              type: integer
              format: int64
            date:
              type: integer
              format: int64
            contentLength:
              type: integer
              format: int64
            contentType:
              $ref: "#/components/schemas/MediaType"
            ifModifiedSince:
              type: integer
              format: int64
            origin:
              type: string
            range:
              type: array
              items:
                $ref: "#/components/schemas/HttpRange"
            allow:
              uniqueItems: true
              type: array
              items:
                $ref: "#/components/schemas/HttpMethod"
            cacheControl:
              type: string
            contentLanguage:
              type: object
              properties:
                language:
                  type: string
                displayName:
                  type: string
                country:
                  type: string
                variant:
                  type: string
                script:
                  type: string
                unicodeLocaleAttributes:
                  uniqueItems: true
                  type: array
                  items:
                    type: string
                unicodeLocaleKeys:
                  uniqueItems: true
                  type: array
                  items:
                    type: string
                displayLanguage:
                  type: string
                displayScript:
                  type: string
                displayCountry:
                  type: string
                displayVariant:
                  type: string
                extensionKeys:
                  uniqueItems: true
                  type: array
                  items:
                    type: string
                iso3Language:
                  type: string
                iso3Country:
                  type: string
            etag:
              type: string
            accessControlAllowOrigin:
              type: string
            accessControlAllowMethods:
              type: array
              items:
                $ref: "#/components/schemas/HttpMethod"
            accessControlAllowHeaders:
              type: array
              items:
                type: string
            accessControlExposeHeaders:
              type: array
              items:
                type: string
            accessControlAllowCredentials:
              type: boolean
            accessControlMaxAge:
              type: integer
              format: int64
            accessControlRequestMethod:
              $ref: "#/components/schemas/HttpMethod"
            accessControlRequestHeaders:
              type: array
              items:
                type: string
            accept:
              type: array
              items:
                $ref: "#/components/schemas/MediaType"
            acceptLanguageAsLocales:
              type: array
              items:
                type: object
                properties:
                  language:
                    type: string
                  displayName:
                    type: string
                  country:
                    type: string
                  variant:
                    type: string
                  script:
                    type: string
                  unicodeLocaleAttributes:
                    uniqueItems: true
                    type: array
                    items:
                      type: string
                  unicodeLocaleKeys:
                    uniqueItems: true
                    type: array
                    items:
                      type: string
                  displayLanguage:
                    type: string
                  displayScript:
                    type: string
                  displayCountry:
                    type: string
                  displayVariant:
                    type: string
                  extensionKeys:
                    uniqueItems: true
                    type: array
                    items:
                      type: string
                  iso3Language:
                    type: string
                  iso3Country:
                    type: string
            acceptPatch:
              type: array
              items:
                $ref: "#/components/schemas/MediaType"
            bearerAuth:
              type: string
              writeOnly: true
            expires:
              type: integer
              format: int64
            ifMatch:
              type: array
              items:
                type: string
            ifNoneMatch:
              type: array
              items:
                type: string
            ifUnmodifiedSince:
              type: integer
              format: int64
            pragma:
              type: string
            upgrade:
              type: string
            vary:
              type: array
              items:
                type: string
            acceptLanguage:
              type: array
              items:
                type: object
                properties:
                  range:
                    type: string
                  weight:
                    type: number
                    format: double
            basicAuth:
              type: string
              writeOnly: true
          additionalProperties:
            type: array
            items:
              type: string
        statusCode:
          $ref: "#/components/schemas/HttpStatusCode"
        typeMessageCode:
          type: string
        detailMessageCode:
          type: string
        detailMessageArguments:
          type: array
          items:
            type: object
        titleMessageCode:
          type: string
    HttpMethod:
      type: object
    HttpRange:
      type: object
    HttpStatusCode:
      type: object
      properties:
        is4xxClientError:
          type: boolean
        error:
          type: boolean
        is5xxServerError:
          type: boolean
        is1xxInformational:
          type: boolean
        is2xxSuccessful:
          type: boolean
        is3xxRedirection:
          type: boolean
    MediaType:
      type: object
      properties:
        type:
          type: string
        subtype:
          type: string
        parameters:
          type: object
          additionalProperties:
            type: string
        qualityValue:
          type: number
          format: double
        wildcardType:
          type: boolean
        wildcardSubtype:
          type: boolean
        subtypeSuffix:
          type: string
        charset:
          type: string
        concrete:
          type: boolean
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        properties:
          type: object
          additionalProperties:
            type: object
...
    UserRegisterDto:
      required:
        - email
        - nickname
        - password
      type: object
      properties:
        nickname:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum:
            - ADMIN
            - USER
    AuthResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nickname:
          type: string
        email:
          type: string
        role:
          type: string
          enum:
            - ADMIN
            - USER
        token:
          type: string
    UserLoginDto:
      required:
        - email
        - password
      type: object
      properties:
        email:
          type: string
        password:
          type: string
...
    UserDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nickname:
          type: string
        email:
          type: string
        role:
          type: string
          enum:
            - ADMIN
            - USER
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT